Create lock schema module.
