+++ openfare-lib/src/lock/mod.rs
@@ -2,21 +2,13 @@ use anyhow::Result;
 
 pub mod payee;
 pub mod plan;
+mod schema;
 pub mod shares;
 
 #[cfg(test)]
 mod tests;
 
 pub static FILE_NAME: &str = "OpenFare.lock";
-pub static SCHEME_VERSION: &str = "1";
-
-lazy_static! {
-    static ref SCHEMA: jsonschema::JSONSchema = {
-        let schema = std::include_str!("schema.json");
-        let schema = serde_json::from_str(schema).expect("serde parsed lock schema");
-        jsonschema::JSONSchema::compile(&schema).expect("compiled lock schema")
-    };
-}
 
 /// A software package's OpenFare lock file (OpenFare.lock).
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
@@ -31,34 +23,17 @@ pub struct Lock {
 }
 
 impl Lock {
-    /// Validate lock against schema.
+    /// Validate lock.
     pub fn validate(&self) -> Result<()> {
         let value = serde_json::to_value(&self)?;
-        let result = SCHEMA.validate(&value);
-        if let Err(errors) = result {
-            let error_string = lockfile_errors_to_string(errors);
-            return Err(anyhow::format_err!(
-                "Invalid lockfile\n".to_owned() + &error_string
-            ));
-        }
-        Ok(())
+        schema::validate(&value)
     }
-}
-
-fn lockfile_errors_to_string(errors: jsonschema::ErrorIterator) -> String {
-    let mut error_string = String::new();
-    error_string += &("----------------------\n");
-    for error in errors {
-        error_string += format!("Validation error: {}\n", error).as_str();
-        error_string += format!("Instance path: {}\n", error.instance_path).as_str();
-    }
-    error_string
 }
 
 impl std::default::Default for Lock {
     fn default() -> Self {
         Self {
-            scheme_version: SCHEME_VERSION.to_string(),
+            scheme_version: schema::VERSION.to_string(),
             plans: plan::Plans::new(),
             payees: payee::Payees::new(),
             shares: None,
