+++ openfare-lib/src/lock/schema/schema.json
